staticfiles/staticfiles/staticfiles/staticfiles/requirements.txt
asgiref==3.5.2
automium==0.2.6
automium-web==0.1.1
autopep8==1.6.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
distlib==0.3.4
Django==4.0.5
django-autoslug==1.9.8
django-tinymce==3.4.0
filelock==3.7.1
Flask==2.1.2
Flask-MySQLdb==1.0.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
legacy==0.1.6
login==0.0.6
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
mysqlclient==2.1.0
numpy==1.22.4
pandas==1.4.2
Pillow==9.1.1
platformdirs==2.5.2
plyer==2.0.0
protobuf==4.21.1
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
scikit-learn==1.1.1
scipy==1.8.1
signup==0.0.1
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
threadpoolctl==3.1.0
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
waitress==2.1.2
Werkzeug==2.1.2
